backend/requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14
django-cors-headers>=4.0
gunicorn>=21.0
whitenoise>=6.0
